Adaranline/requirements.txt
Flask==3.1.2
mysql-connector-python==8.4.0
PyJWT==2.11.0
requests==2.32.5
user-agents==2.2.0
Werkzeug==3.1.5
Flask-Mail==0.9.1
razorpay==1.4.2
gunicorn==21.2.0